requirements.txt
langchain==0.3.26
langchain_community==0.3.27
langchain_core==0.3.68
langchain_groq==0.3.5
numpy==2.3.1
pinecone==7.0.2
python-dotenv==1.1.1
sentence_transformers==5.0.0
streamlit==1.46.1